requirements.txt
Django==2.0.1
django-import-export==0.7.0
